delete old files and change name of new ochestrator file.
